# Expose ReferenceResolver and rename ReferenceHandling to ReferenceHandler