… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Collection/IEnumerableDefaultConverter.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Object/ObjectDefaultConverter.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Object/ObjectWithParameterizedConstructorConverter.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/DefaultReferenceResolver.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Read.HandleMetadata.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Read.HandlePropertyName.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Write.HandleMetadata.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializerOptions.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/PreserveReferenceHandler.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/PreserveReferenceResolver.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/ReadStack.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/ReferenceHandler.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/ReferenceHandlerOfT.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/ReferenceResolver.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/WriteStack.cs
… src/libraries/System.Text.Json/src/System/Text/Json/ThrowHelper.Serialization.cs
… src/libraries/System.Text.Json/tests/Serialization/ConstructorTests/ConstructorTests.Exceptions.cs
… src/libraries/System.Text.Json/tests/Serialization/OptionsTests.cs
… src/libraries/System.Text.Json/tests/Serialization/ReferenceHandlerTests.Deserialize.cs
… src/libraries/System.Text.Json/tests/Serialization/ReferenceHandlerTests.Serialize.cs
… src/libraries/System.Text.Json/tests/Serialization/ReferenceHandlerTests.cs
… src/libraries/System.Text.Json/tests/Serialization/Stream.Collections.cs
… src/libraries/System.Text.Json/tests/System.Text.Json.Tests.csproj