Address the following as it was detailed in https://github.com/dotnet/runtime/issues/30820#issuecomment-625444164:
* The naming changes for `ReferenceHandling` to resemble a class.
* Exposure of the `ReferenceResolver` class.
* Add `ReferenceHandler<T> where T : ReferenceResolver` to allow users to safely provide their own resolver implementation that will be instantiated on each call to (De)Serialize. 
* Make `ReferenceHandler` abstract so if you override it you can opt-in for using a persistent resolver.

This PR effectively closes https://github.com/dotnet/runtime/issues/30820 as no more changes would be pending to address.