@Jozkee 2020-05-29
This change does not regress performance on the out-of-box path.

## Read 

**Before:**
|                                                Type |                   Method |         Mean |       Error |      StdDev |       Median |          Min |          Max |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|---------------------------------------------------- |------------------------- |-------------:|------------:|------------:|-------------:|-------------:|-------------:|--------:|-------:|------:|----------:|
|                                 ReadJson&lt;ArrayList&gt; | DeserializeFromUtf8Bytes |  50,628.0 ns |   238.56 ns |   199.21 ns |  50,577.9 ns |  50,413.1 ns |  51,149.3 ns |  4.8761 | 0.6095 |     - |   31016 B |
|                                ReadJson&lt;BinaryData&gt; | DeserializeFromUtf8Bytes |     522.8 ns |     4.57 ns |     4.28 ns |     522.0 ns |     514.8 ns |     530.4 ns |  0.1689 |      - |     - |    1072 B |
|                ReadJson&lt;Dictionary&lt;String, String&gt;&gt; | DeserializeFromUtf8Bytes |  20,494.8 ns |   123.63 ns |   115.65 ns |  20,442.9 ns |  20,364.6 ns |  20,738.1 ns |  4.0690 | 0.5697 |     - |   25872 B |
|                           ReadJson&lt;HashSet&lt;String&gt;&gt; | DeserializeFromUtf8Bytes |  13,761.3 ns |   125.61 ns |   117.50 ns |  13,762.9 ns |  13,601.4 ns |  14,015.2 ns |  2.3997 | 0.1636 |     - |   15216 B |
|                                 ReadJson&lt;Hashtable&gt; | DeserializeFromUtf8Bytes |  67,377.3 ns | 1,659.28 ns | 1,910.83 ns |  67,248.2 ns |  64,903.0 ns |  72,845.5 ns |  7.0404 | 1.3038 |     - |   45282 B |
|       ReadJson&lt;ImmutableDictionary&lt;String, String&gt;&gt; | DeserializeFromUtf8Bytes |  46,786.9 ns |   379.34 ns |   354.84 ns |  46,728.0 ns |  46,191.5 ns |  47,358.7 ns |  4.9926 | 0.5547 |     - |   32368 B |
| ReadJson&lt;ImmutableSortedDictionary&lt;String, String&gt;&gt; | DeserializeFromUtf8Bytes |  72,652.2 ns |   733.89 ns |   686.48 ns |  72,548.6 ns |  71,654.5 ns |  73,647.9 ns |  6.2500 | 0.8523 |     - |   39208 B |
|                            ReadJson&lt;IndexViewModel&gt; | DeserializeFromUtf8Bytes |  29,374.9 ns |   331.78 ns |   294.11 ns |  29,270.6 ns |  29,068.4 ns |  29,973.3 ns |  3.4961 | 0.3496 |     - |   22504 B |
|                                     ReadJson&lt;Int32&gt; | DeserializeFromUtf8Bytes |     106.3 ns |     0.89 ns |     0.83 ns |     106.4 ns |     104.7 ns |     107.3 ns |       - |      - |     - |         - |
|                 ReadJson&lt;LargeStructWithProperties&gt; | DeserializeFromUtf8Bytes |   1,130.5 ns |     6.80 ns |     6.36 ns |   1,129.6 ns |   1,122.0 ns |   1,141.1 ns |  0.0274 |      - |     - |     200 B |
|                                  ReadJson&lt;Location&gt; | DeserializeFromUtf8Bytes |   1,181.6 ns |    10.60 ns |     9.39 ns |   1,180.3 ns |   1,165.8 ns |   1,197.6 ns |  0.0683 |      - |     - |     448 B |
|                            ReadJson&lt;LoginViewModel&gt; | DeserializeFromUtf8Bytes |     472.2 ns |     9.76 ns |    10.44 ns |     472.9 ns |     459.8 ns |     496.2 ns |  0.0264 |      - |     - |     168 B |
|                   ReadJson&lt;MyEventsListerViewModel&gt; | DeserializeFromUtf8Bytes | 322,115.1 ns | 2,461.36 ns | 2,302.36 ns | 322,283.0 ns | 318,197.7 ns | 326,914.3 ns | 11.4213 | 2.5381 |     - |   78714 B |
|                ReadJson&lt;SimpleStructWithProperties&gt; | DeserializeFromUtf8Bytes |     346.9 ns |    20.37 ns |    23.45 ns |     341.9 ns |     318.9 ns |     394.5 ns |  0.0090 |      - |     - |      64 B |
|                                 ReadJson&lt;ArrayList&gt; |    DeserializeFromStream |  49,650.4 ns |   642.03 ns |   600.55 ns |  49,468.5 ns |  49,076.7 ns |  51,492.8 ns |  4.9300 | 0.5916 |     - |   31088 B |
|                                ReadJson&lt;BinaryData&gt; |    DeserializeFromStream |     865.7 ns |    11.55 ns |    10.80 ns |     868.0 ns |     842.0 ns |     879.1 ns |  0.1807 |      - |     - |    1144 B |
|                ReadJson&lt;Dictionary&lt;String, String&gt;&gt; |    DeserializeFromStream |  22,967.5 ns |   258.54 ns |   241.84 ns |  23,031.3 ns |  22,565.0 ns |  23,358.8 ns |  4.1071 | 0.4464 |     - |   25944 B |
|                           ReadJson&lt;HashSet&lt;String&gt;&gt; |    DeserializeFromStream |  14,812.5 ns |   130.13 ns |   115.36 ns |  14,803.2 ns |  14,637.5 ns |  15,067.6 ns |  2.4152 | 0.1767 |     - |   15288 B |
|                                 ReadJson&lt;Hashtable&gt; |    DeserializeFromStream |  66,574.3 ns | 1,256.13 ns | 1,233.69 ns |  65,985.9 ns |  65,664.9 ns |  69,532.9 ns |  7.1694 | 1.0621 |     - |   45352 B |
|       ReadJson&lt;ImmutableDictionary&lt;String, String&gt;&gt; |    DeserializeFromStream |  49,008.3 ns |   817.75 ns |   682.86 ns |  48,836.5 ns |  48,221.3 ns |  50,290.3 ns |  5.0310 | 0.5805 |     - |   32440 B |
| ReadJson&lt;ImmutableSortedDictionary&lt;String, String&gt;&gt; |    DeserializeFromStream |  81,203.2 ns |   489.68 ns |   458.04 ns |  81,177.6 ns |  80,337.4 ns |  81,922.0 ns |  6.1528 | 0.6477 |     - |   39280 B |
|                            ReadJson&lt;IndexViewModel&gt; |    DeserializeFromStream |  32,711.6 ns |   445.74 ns |   416.94 ns |  32,599.6 ns |  32,067.0 ns |  33,552.4 ns |  3.5220 | 0.3913 |     - |   22576 B |
|                                     ReadJson&lt;Int32&gt; |    DeserializeFromStream |     303.7 ns |     1.62 ns |     1.51 ns |     303.4 ns |     301.8 ns |     306.3 ns |  0.0111 |      - |     - |      72 B |
|                 ReadJson&lt;LargeStructWithProperties&gt; |    DeserializeFromStream |   1,624.8 ns |    15.91 ns |    14.88 ns |   1,624.9 ns |   1,590.6 ns |   1,646.4 ns |  0.0516 |      - |     - |     328 B |
|                                  ReadJson&lt;Location&gt; |    DeserializeFromStream |   1,612.5 ns |    30.67 ns |    35.32 ns |   1,593.9 ns |   1,572.3 ns |   1,685.7 ns |  0.0819 |      - |     - |     520 B |
|                            ReadJson&lt;LoginViewModel&gt; |    DeserializeFromStream |     762.4 ns |    15.22 ns |    16.29 ns |     760.6 ns |     742.5 ns |     802.2 ns |  0.0357 |      - |     - |     240 B |
|                   ReadJson&lt;MyEventsListerViewModel&gt; |    DeserializeFromStream | 384,195.9 ns | 3,742.68 ns | 3,500.90 ns | 384,658.1 ns | 377,496.5 ns | 388,682.9 ns | 12.1951 | 3.0488 |     - |   79514 B |
|                ReadJson&lt;SimpleStructWithProperties&gt; |    DeserializeFromStream |     580.4 ns |    11.06 ns |    10.35 ns |     577.9 ns |     569.0 ns |     606.2 ns |  0.0207 |      - |     - |     144 B |

**Now:**
|                                                Type |                   Method |         Mean |       Error |      StdDev |       Median |          Min |          Max |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|---------------------------------------------------- |------------------------- |-------------:|------------:|------------:|-------------:|-------------:|-------------:|--------:|-------:|------:|----------:|
|                                 ReadJson&lt;ArrayList&gt; | DeserializeFromUtf8Bytes |  48,300.3 ns |   189.82 ns |   177.56 ns |  48,240.3 ns |  48,113.4 ns |  48,674.2 ns |  4.8022 | 0.5763 |     - |   31016 B |
|                                ReadJson&lt;BinaryData&gt; | DeserializeFromUtf8Bytes |     507.1 ns |     3.49 ns |     3.10 ns |     506.8 ns |     502.6 ns |     513.1 ns |  0.1689 |      - |     - |    1072 B |
|                ReadJson&lt;Dictionary&lt;String, String&gt;&gt; | DeserializeFromUtf8Bytes |  21,058.2 ns |   266.56 ns |   249.34 ns |  21,029.9 ns |  20,728.5 ns |  21,441.8 ns |  4.0833 | 0.5833 |     - |   25872 B |
|                           ReadJson&lt;HashSet&lt;String&gt;&gt; | DeserializeFromUtf8Bytes |  14,006.4 ns |   161.10 ns |   142.82 ns |  14,036.1 ns |  13,768.6 ns |  14,234.0 ns |  2.4186 | 0.1649 |     - |   15216 B |
|                                 ReadJson&lt;Hashtable&gt; | DeserializeFromUtf8Bytes |  65,265.2 ns |   282.61 ns |   250.53 ns |  65,318.3 ns |  64,776.7 ns |  65,632.7 ns |  7.0922 | 1.3134 |     - |   45280 B |
|       ReadJson&lt;ImmutableDictionary&lt;String, String&gt;&gt; | DeserializeFromUtf8Bytes |  46,677.9 ns |   417.39 ns |   390.42 ns |  46,621.7 ns |  46,143.0 ns |  47,459.3 ns |  4.9779 | 0.5531 |     - |   32368 B |
| ReadJson&lt;ImmutableSortedDictionary&lt;String, String&gt;&gt; | DeserializeFromUtf8Bytes |  77,860.1 ns |   498.28 ns |   416.09 ns |  77,809.6 ns |  77,239.0 ns |  78,895.8 ns |  6.1881 | 0.9282 |     - |   39208 B |
|                            ReadJson&lt;IndexViewModel&gt; | DeserializeFromUtf8Bytes |  29,217.9 ns |   235.41 ns |   220.20 ns |  29,218.2 ns |  28,812.0 ns |  29,689.7 ns |  3.4961 | 0.3496 |     - |   22504 B |
|                                     ReadJson&lt;Int32&gt; | DeserializeFromUtf8Bytes |     104.8 ns |     0.52 ns |     0.46 ns |     104.7 ns |     104.1 ns |     105.5 ns |       - |      - |     - |         - |
|                 ReadJson&lt;LargeStructWithProperties&gt; | DeserializeFromUtf8Bytes |   1,164.1 ns |     7.07 ns |     6.27 ns |   1,161.5 ns |   1,156.4 ns |   1,176.9 ns |  0.0278 |      - |     - |     200 B |
|                                  ReadJson&lt;Location&gt; | DeserializeFromUtf8Bytes |   1,194.4 ns |     9.27 ns |     7.74 ns |   1,197.2 ns |   1,173.8 ns |   1,201.6 ns |  0.0674 |      - |     - |     448 B |
|                            ReadJson&lt;LoginViewModel&gt; | DeserializeFromUtf8Bytes |     455.1 ns |     6.87 ns |     6.09 ns |     455.5 ns |     446.2 ns |     465.6 ns |  0.0255 |      - |     - |     168 B |
|                   ReadJson&lt;MyEventsListerViewModel&gt; | DeserializeFromUtf8Bytes | 326,594.4 ns | 2,843.42 ns | 2,520.62 ns | 326,530.1 ns | 321,413.7 ns | 331,308.2 ns | 11.7801 | 2.6178 |     - |   78714 B |
|                ReadJson&lt;SimpleStructWithProperties&gt; | DeserializeFromUtf8Bytes |     324.4 ns |     1.98 ns |     1.85 ns |     324.4 ns |     320.8 ns |     327.2 ns |  0.0091 |      - |     - |      64 B |
|                                 ReadJson&lt;ArrayList&gt; |    DeserializeFromStream |  52,083.5 ns |   418.37 ns |   370.87 ns |  52,218.8 ns |  51,274.4 ns |  52,500.3 ns |  4.7827 | 0.6238 |     - |   31090 B |
|                                ReadJson&lt;BinaryData&gt; |    DeserializeFromStream |     844.1 ns |     9.87 ns |     9.24 ns |     846.1 ns |     824.4 ns |     852.7 ns |  0.1813 |      - |     - |    1144 B |
|                ReadJson&lt;Dictionary&lt;String, String&gt;&gt; |    DeserializeFromStream |  22,176.6 ns |   232.96 ns |   217.91 ns |  22,113.5 ns |  21,935.3 ns |  22,521.4 ns |  4.0607 | 0.4414 |     - |   25944 B |
|                           ReadJson&lt;HashSet&lt;String&gt;&gt; |    DeserializeFromStream |  14,867.0 ns |   138.48 ns |   129.53 ns |  14,813.8 ns |  14,709.5 ns |  15,096.5 ns |  2.4243 | 0.1774 |     - |   15288 B |
|                                 ReadJson&lt;Hashtable&gt; |    DeserializeFromStream |  66,353.0 ns |   464.60 ns |   434.59 ns |  66,439.5 ns |  65,641.0 ns |  66,991.6 ns |  7.1828 | 1.0641 |     - |   45352 B |
|       ReadJson&lt;ImmutableDictionary&lt;String, String&gt;&gt; |    DeserializeFromStream |  50,086.9 ns |   460.06 ns |   430.34 ns |  50,057.0 ns |  49,222.7 ns |  50,798.6 ns |  5.0896 | 0.6107 |     - |   32440 B |
| ReadJson&lt;ImmutableSortedDictionary&lt;String, String&gt;&gt; |    DeserializeFromStream |  76,215.1 ns |   724.63 ns |   677.82 ns |  76,110.2 ns |  75,414.1 ns |  77,506.5 ns |  6.1576 | 0.6158 |     - |   39280 B |
|                            ReadJson&lt;IndexViewModel&gt; |    DeserializeFromStream |  32,288.7 ns |   313.06 ns |   292.84 ns |  32,270.4 ns |  31,832.4 ns |  32,740.0 ns |  3.4709 | 0.3857 |     - |   22576 B |
|                                     ReadJson&lt;Int32&gt; |    DeserializeFromStream |     300.0 ns |     3.40 ns |     3.18 ns |     298.3 ns |     296.4 ns |     305.5 ns |  0.0107 |      - |     - |      72 B |
|                 ReadJson&lt;LargeStructWithProperties&gt; |    DeserializeFromStream |   1,671.6 ns |    14.93 ns |    13.23 ns |   1,671.6 ns |   1,648.6 ns |   1,700.8 ns |  0.0463 |      - |     - |     328 B |
|                                  ReadJson&lt;Location&gt; |    DeserializeFromStream |   1,593.5 ns |    14.26 ns |    13.34 ns |   1,593.2 ns |   1,570.1 ns |   1,612.3 ns |  0.0823 |      - |     - |     520 B |
|                            ReadJson&lt;LoginViewModel&gt; |    DeserializeFromStream |     753.0 ns |     8.09 ns |     7.57 ns |     754.8 ns |     743.3 ns |     770.6 ns |  0.0361 |      - |     - |     240 B |
|                   ReadJson&lt;MyEventsListerViewModel&gt; |    DeserializeFromStream | 377,660.1 ns | 4,094.25 ns | 3,829.76 ns | 378,626.5 ns | 372,579.7 ns | 384,137.9 ns | 12.2324 | 3.0581 |     - |   79515 B |
|                ReadJson&lt;SimpleStructWithProperties&gt; |    DeserializeFromStream |     558.0 ns |     4.21 ns |     3.94 ns |     558.4 ns |     552.0 ns |     564.1 ns |  0.0222 |      - |     - |     144 B |


| before/now \(mean\)   |
|--------------|
| 0\.954023465 |
| 0\.969969396 |
| 1\.0274899   |
| 1\.017810817 |
| 0\.968652647 |
| 0\.997670288 |
| 1\.071682619 |
| 0\.994655301 |
| 0\.985888993 |
| 1\.029721362 |
| 1\.010832769 |
| 0\.963786531 |
| 1\.013905899 |
| 0\.93513981  |
| 1\.04900464  |
| 0\.975049093 |
| 0\.965564384 |
| 1\.003679325 |
| 0\.996675894 |
| 1\.022008517 |
| 0\.938572618 |
| 0\.987071864 |
| 0\.987816925 |
| 1\.028803545 |
| 0\.988217054 |
| 0\.987670514 |
| 0\.982988366 |
| 0\.961405927 |

## Write

**Before:**
|                               Type |               Method |         Mean |       Error |      StdDev |       Median |          Min |          Max |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|----------------------------------- |--------------------- |-------------:|------------:|------------:|-------------:|-------------:|-------------:|--------:|-------:|------:|----------:|
|          WriteJson&lt;IndexViewModel&gt; | SerializeToUtf8Bytes |  17,090.9 ns |   124.41 ns |   116.37 ns |  17,089.0 ns |  16,954.5 ns |  17,323.3 ns |  2.0650 | 0.0688 |     - |   13016 B |
|                   WriteJson&lt;Int32&gt; | SerializeToUtf8Bytes |     144.6 ns |     1.48 ns |     1.39 ns |     145.1 ns |     141.2 ns |     145.9 ns |  0.0290 |      - |     - |     184 B |
|                WriteJson&lt;Location&gt; | SerializeToUtf8Bytes |     843.9 ns |    10.49 ns |     9.82 ns |     849.4 ns |     824.7 ns |     855.3 ns |  0.0600 |      - |     - |     384 B |
|          WriteJson&lt;LoginViewModel&gt; | SerializeToUtf8Bytes |     343.4 ns |     5.18 ns |     4.85 ns |     342.5 ns |     335.8 ns |     352.9 ns |  0.0421 |      - |     - |     264 B |
| WriteJson&lt;MyEventsListerViewModel&gt; | SerializeToUtf8Bytes | 431,276.0 ns | 2,504.57 ns | 2,342.78 ns | 431,172.6 ns | 426,409.1 ns | 434,945.3 ns | 28.7162 | 5.0676 |     - |  191432 B |
|          WriteJson&lt;IndexViewModel&gt; |    SerializeToStream |  16,731.5 ns |   183.90 ns |   172.02 ns |  16,755.3 ns |  16,509.8 ns |  16,959.0 ns |  0.0667 |      - |     - |     432 B |
|                   WriteJson&lt;Int32&gt; |    SerializeToStream |     226.1 ns |     1.63 ns |     1.52 ns |     225.7 ns |     223.7 ns |     229.3 ns |  0.0240 |      - |     - |     152 B |
|                WriteJson&lt;Location&gt; |    SerializeToStream |     940.7 ns |    11.58 ns |    10.83 ns |     944.3 ns |     923.8 ns |     954.1 ns |  0.0229 |      - |     - |     152 B |
|          WriteJson&lt;LoginViewModel&gt; |    SerializeToStream |     430.9 ns |     2.42 ns |     2.02 ns |     431.4 ns |     426.7 ns |     433.1 ns |  0.0227 |      - |     - |     152 B |
| WriteJson&lt;MyEventsListerViewModel&gt; |    SerializeToStream | 415,983.5 ns | 3,423.78 ns | 3,202.61 ns | 414,854.8 ns | 412,741.9 ns | 422,329.6 ns | 18.0921 |      - |     - |  116104 B |

**Now:**
|                               Type |               Method |         Mean |       Error |      StdDev |       Median |          Min |          Max |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|----------------------------------- |--------------------- |-------------:|------------:|------------:|-------------:|-------------:|-------------:|--------:|-------:|------:|----------:|
|          WriteJson&lt;IndexViewModel&gt; | SerializeToUtf8Bytes |  16,909.1 ns |    62.43 ns |    52.13 ns |  16,926.6 ns |  16,832.1 ns |  17,002.1 ns |  2.0336 | 0.0678 |     - |   13016 B |
|                   WriteJson&lt;Int32&gt; | SerializeToUtf8Bytes |     141.1 ns |     1.11 ns |     0.98 ns |     141.1 ns |     139.2 ns |     142.9 ns |  0.0289 |      - |     - |     184 B |
|                WriteJson&lt;Location&gt; | SerializeToUtf8Bytes |     883.7 ns |     8.48 ns |     7.93 ns |     885.5 ns |     868.6 ns |     894.1 ns |  0.0608 |      - |     - |     384 B |
|          WriteJson&lt;LoginViewModel&gt; | SerializeToUtf8Bytes |     347.4 ns |     8.39 ns |     9.66 ns |     347.6 ns |     332.0 ns |     367.8 ns |  0.0408 |      - |     - |     264 B |
| WriteJson&lt;MyEventsListerViewModel&gt; | SerializeToUtf8Bytes | 421,477.8 ns | 5,711.71 ns | 5,342.74 ns | 421,054.7 ns | 414,107.3 ns | 430,927.0 ns | 28.7162 | 5.0676 |     - |  191432 B |
|          WriteJson&lt;IndexViewModel&gt; |    SerializeToStream |  17,151.0 ns |   196.95 ns |   184.22 ns |  17,208.2 ns |  16,784.8 ns |  17,387.3 ns |  0.0670 |      - |     - |     432 B |
|                   WriteJson&lt;Int32&gt; |    SerializeToStream |     214.7 ns |     2.22 ns |     2.08 ns |     215.0 ns |     211.8 ns |     218.3 ns |  0.0239 |      - |     - |     152 B |
|                WriteJson&lt;Location&gt; |    SerializeToStream |     922.5 ns |     8.25 ns |     7.31 ns |     922.2 ns |     912.4 ns |     936.1 ns |  0.0203 |      - |     - |     152 B |
|          WriteJson&lt;LoginViewModel&gt; |    SerializeToStream |     437.0 ns |     2.23 ns |     2.09 ns |     436.5 ns |     435.0 ns |     441.0 ns |  0.0227 |      - |     - |     152 B |
| WriteJson&lt;MyEventsListerViewModel&gt; |    SerializeToStream | 408,500.2 ns | 2,773.06 ns | 2,458.25 ns | 408,301.8 ns | 405,188.2 ns | 413,476.8 ns | 18.0921 |      - |     - |  116104 B |

| before/now \(mean\) |
|---------------------|
| 0\.98936276         |
| 0\.975795297        |
| 1\.047161986        |
| 1\.011648224        |
| 0\.977280906        |
| 1\.025072468        |
| 0\.949579832        |
| 0\.980652705        |
| 1\.014156417        |
| 0\.982010585        |